Fix int/size_t conversion.

+++ cpp/models/abm/world.cpp
@@ -160,10 +160,10 @@ Location& World::find_location(LocationType type, const Person& person)
 size_t World::get_subpopulation_combined(TimePoint t, InfectionState s, LocationType type) const
 {
     auto& locs = m_locations[(uint32_t)type];
-    return (size_t)std::accumulate(locs.begin(), locs.end(), 0,
-                                   [&](size_t running_sum, const std::unique_ptr<Location>& loc) {
-                                       return running_sum + loc->get_subpopulation(t, s);
-                                   });
+    return std::accumulate(locs.begin(), locs.end(), 0UL,
+                           [&](size_t running_sum, const std::unique_ptr<Location>& loc) {
+                               return running_sum + loc->get_subpopulation(t, s);
+                           });
 }
 
 MigrationParameters& World::get_migration_parameters()
